requirements.txt
Flask==2.3.2
Flask-Cors==4.0.0
mysql-connector-python==8.0.33
mysqlclient==2.1.1  
PyJWT==2.7.0
